Added the Bar System, Character System, and ATB Manager
The Character System (Incomplete) embodies the character as it's own ScriptableObject so that characters can quickly be made using Create/Character. Allowing for swift assembly of characters, manual, or automatic (for future reference)!

The ATB Manager is a timing system that will be read as part of the overall combat system, it fuses well with the ATBBar which is part of the BarSystem, a system.... made for bars...