Rigged and IK'd Millie Basic Asset
Rigged Millie, changed a few mesh strengths for IK, and added IK to sets of bones that have more than 1 bone in the chain.

Moved a few files around, updated character core to be accepted in Animation Event Systems.

Added new tag layers for different renderings.

Removed old Animator off of Millie.

Will add proper animation combats in the future... so that Sean can be defeated once and for all.