Attempting to create the Action Rail System
The action rail system is a dedicated system that is designed to create a micro-timed environment that allows players to make adjustements to the character's attacks.

Unfortunately, this version currently doesn't work due to technical difficulties. and self-confusion pertaining to the type of interface I want in the Action Rail System q -p